add map for sub level collection
